+++ source/2_platform/rex_directx/include/rex_directx/resources/dx_upload_buffer.h
@@ -17,14 +17,14 @@ namespace rex
       UploadBuffer(const wrl::ComPtr<ID3D12Resource>& uploadBuffer, D3D12_RESOURCE_STATES startState = D3D12_RESOURCE_STATE_COMMON);
       ~UploadBuffer();
 
-      void write(CommandList* cmdList, Resource* dstResource, const void* data, s32 size);
+      void write(CommandList* cmdList, Resource* dstResource, const void* data, s64 size);
 
       void reset();
 
     private:
       wrl::ComPtr<ID3D12Resource> m_upload_buffer;
       void* m_mapped_data;
-      s32 m_offset;
+      s64 m_offset;
       D3D12_RESOURCE_STATES m_resource_state;
     };
   }
